- link handling changes, update the dupe type when right clicking on a magnet/release name - open dupe files in links from disk - fix "Don't share files bigger than x MB" when the limit is more than 2048MB
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@1271 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ ShareManager.cpp
@@ -2809,7 +2809,7 @@ bool ShareManager::checkSharedName(const string& aPath, bool isDir, bool report 
 		if(SETTING(NO_ZERO_BYTE) && !(size > 0))
 			return false;
 
-		if ((SETTING(MAX_FILE_SIZE_SHARED) != 0) && (size > (SETTING(MAX_FILE_SIZE_SHARED)*1024*1024))) {
+		if ((SETTING(MAX_FILE_SIZE_SHARED) != 0) && (size > ((int64_t)SETTING(MAX_FILE_SIZE_SHARED)*1024*1024))) {
 			if (report) {
 				LogManager::getInstance()->message(STRING(BIG_FILE_NOT_SHARED) + " " + aPath, LogManager::LOG_INFO);
 			}
